% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtw.R
\name{get_dtw_dist_bidirectional}
\alias{get_dtw_dist_bidirectional}
\title{Bidirectional DTW Distance}
\usage{
get_dtw_dist_bidirectional(x, y)
}
\arguments{
\item{x}{list of x input}

\item{y}{list of y input}
}
\value{
numeric
}
\description{
Get bidirectional DTW distance.
}
\examples{
get_dtw_dist_bidirectional(c(1:1000),c(1:1000))

}
