% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pGRNDB}
\alias{pGRNDB}
\title{pGRN example data}
\format{
\subsection{\code{pGRNDB}}{

A list with items expression and ptime
\describe{
\item{expression}{data frame of single cell expression}
\item{ptime}{pseudotime of the single cells}
...
}
}
}
\source{
pGRN
}
\usage{
pGRNDB
}
\description{
A list with expression dataframe and pseudotime dataframe
}
\keyword{datasets}
