% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtw.R
\name{slideWindows}
\alias{slideWindows}
\title{Sliding Window Average}
\usage{
slideWindows(data, window = 2, step = 1)
}
\arguments{
\item{data}{list of expression}

\item{window}{sliding window size}

\item{step}{sliding window step size}
}
\value{
list/vector of sliding windows with average expression value
}
\description{
Get sliding windows average values for given vector/list
}
\examples{
slideWindows(c(1:1000),window=200,step=100)
slideWindows(c(1:1000),window=100,step=50)

}
